上激光器调试成功
修正更改参数后激光器重新启动电压不连续的问题
锁相滤波器问题修正

+++ Sources_Bsp/Bsp_AD5546.h
@@ -13,7 +13,7 @@ extern "C" {
 
 FP32 Bsp_AD5546HexToVolt(INT16U hex);
 
-INT16U Bsp_AD5546VoltToHex(FP64 volt);
+INT16U Bsp_AD5546VoltToHex(FP32 volt);
 
 void Bsp_AD5546Disable(void);
